### 2021-05-24 #3310 @whedon
[PRE REVIEW]: PyAFBF: a Python library to sample image textures from the anisotropic fractional Brownian field.
**Submitting author:** @fjprichard (<a href="http://orcid.org/0000-0001-5146-9894">Frédéric Richard</a>)
**Repository:** <a href="https://github.com/fjprichard/PyAFBF/" target ="_blank">https://github.com/fjprichard/PyAFBF/</a>
**Version:** v0.0.2
