+++ docs/_sources/index.rst.txt
@@ -20,13 +20,15 @@ Package features
 
 - Random definition of simulated fields,
 
-- Extensions to other fields (deformed fields, intrinsic fields, heterogeneous fields, binary patterns).
+- Extensions to related fields (deformed fields, intrinsic fields, heterogeneous fields, binary patterns).
 
 
 Installation from sources
 =========================
 
-In the root directory of the package, just do
+The package source can be downloaded from the `repository <https://github.com/fjprichard/PyAFBF>`_. 
+
+To install the package, write the following commands in the root directory where the package was downloaded:
 
 .. code-block:: python
 
@@ -34,16 +36,23 @@ In the root directory of the package, just do
     or
     pip install -e .
 
+After installation, the package can be tested by running 
+
+.. code-block:: python
+
+    python tests/TestAll.py
+
+If no output appears, then the package is correctly installed.
+
 Communication to the author
 ===========================
 
-PyAFBF is developed and maintained by Frédéric Richard. For feed-back, contributions, bug reports, contact directly the author at
-<frederic.richard_at_univ-amu.fr>.
+PyAFBF is developed and maintained by Frédéric Richard. For feed-back, contributions, bug reports, contact directly the `author <https://github.com/fjprichard>`_, or use the `discussion <https://github.com/fjprichard/PyAFBF/discussions>`_ facility.
 
 Citation
 ========
 
-When using PyAFBF, please cite the original paper :cite:`Bierme-2015-TBM`, and the companion paper. 
+When using PyAFBF, please cite the original paper :cite:`Bierme-2015-TBM`.
 
 Licence
 =======
