+++ joss/refs.bib
@@ -49,7 +49,8 @@ edition = {2nd}
   number={7},
   pages={683--691},
   year={2004},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi="10.1016/j.cageo.2004.03.012"
 }
 
 @Article{Schlater-2015,
@@ -70,7 +71,8 @@ edition = {2nd}
   volume={23},
   number={1},
   pages={1--23},
-  year={2007}
+  year={2007},
+  doi="10.18637/jss.v023.i01"
 }
 
 @Article{Bierme08-ESAIM,
@@ -81,7 +83,7 @@ edition = {2nd}
   volume = 	 {12},
   number = 	 {1},
   pages ={30--50},
-  notes = {},
+  doi="10.1051/ps:2007031"
 }
 
 
@@ -92,6 +94,7 @@ edition = {2nd}
   year = 	 {2017},
  pages = 	 {1--14},
   month = 	 {oct},
+  doi="10.1007/s11222-017-9785-z"
 }
 
 @Article{Richard-2016,
@@ -102,6 +105,7 @@ edition = {2nd}
   volume = 	 {18},
   optnumber = 	 {},
   pages = 	 {147--162},
+  doi="10.1016/j.spasta.2016.02.001"
 }
 @Article{Richard-2015b,
   author = 	 {Richard, F.J.P.},
@@ -110,6 +114,7 @@ edition = {2nd}
   volume ={27},
   pages ={16--20},
   year = 	 {2015},
+  doi="10.1016/j.proenv.2015.07.100"
 }
 
 @Article{Richard-2015,
@@ -120,6 +125,7 @@ edition = {2nd}
   volume = 	 {26},
   number = 	 {3},
   pages = 	 {1279--1304},
+  doi = "10.5705/ss.202014.0077"
 }
 
 @Article{Richard-2010,
@@ -130,6 +136,7 @@ edition = {2nd}
   volume = 	 {36},
   number = 	 {3},
   pages = 	 {227--240},
+  doi="10.1007/s10851-009-0181-y"
 }
 
 @InCollection{Bierme10-Springer,
@@ -144,6 +151,7 @@ edition = {2nd}
   editor = 	 {M. Bergounioux},
   OPTvolume = 	 {},
   OPTnumber = 	 {},
+  doi="10.1007/978-3-642-19604-1_3"
 }
 
 @InCollection{Bierme09-ESAIM,
@@ -155,6 +163,7 @@ edition = {2nd}
   editor = 	 {H. Ammari},
   volume = 	 {26},
   series = 	 {{ESAIM Proceedings}},
+  doi="10.1051/proc/2009008"
 }
 
 
@@ -166,6 +175,7 @@ year = {2015},
 volume = {24},
 number = {3},
 pages = {885--904},
+doi="10.1080/10618600.2014.946603"
 }
 
 @article{Vu2020,
@@ -176,6 +186,7 @@ volume=130,
 number=8, 
 pages={4667--4692}, 
 year=2020,
+doi="10.1016/j.spa.2020.01.012"
 }
 
 @article{Guyon2000,
