+++ joss/refs.bib
@@ -196,7 +196,8 @@ doi="10.1016/j.spa.2020.01.012"
   volume={47},
   number={3},
   pages={307--316},
-  year={2000}
+  year={2000},
+  doi="10.1016/s0167-7152(99)00171-6"
 }
 
 @article{Wood1994,
@@ -215,9 +216,18 @@ doi="10.1016/j.spa.2020.01.012"
   booktitle= {International {C}onference on {I}mage {P}rocessing (ICIP), 2014 {IEEE}},
   pages={6091--6095},
   year={2014},
-  OPTorganization={IEEE}
+  OPTorganization={IEEE},
+  doi="10.1109/icip.2014.7026229"
 }
 
+@ARTICLE{Bonami2003,
+author="Bonami, A. and Estrade, A.",
+title="Anisotropic analysis of some {G}aussian models",
+journal="Journal of {F}ourier {A}nalysis and [A}pplications",
+volume="9",
+pages="215-236",
+year="2003",
+doi="10.1007/s00041-003-0012-2"}
 
 @ARTICLE{Benassi97,
 author="Benassi, A. and Jaffard, S. and Roux, D.",
