### 2021-06-01 2e38d82
Update refs.bib
suppression of a doublon
