+++ joss/refs.bib
@@ -220,15 +220,6 @@ doi="10.1016/j.spa.2020.01.012"
   doi="10.1109/icip.2014.7026229"
 }
 
-@ARTICLE{Bonami2003,
-author="Bonami, A. and Estrade, A.",
-title="Anisotropic analysis of some {G}aussian models",
-journal="Journal of {F}ourier {A}nalysis and [A}pplications",
-volume="9",
-pages="215-236",
-year="2003",
-doi="10.1007/s00041-003-0012-2"}
-
 @ARTICLE{Benassi97,
 author="Benassi, A. and Jaffard, S. and Roux, D.",
 title="{E}lliptic {G}aussian random processes",
@@ -254,7 +245,8 @@ title="Anisotropic analysis of some {G}aussian models",
 journal={{J}ournal of {F}ourier {A}nalysis and {A}pplications},
 volume="9",
 pages="215-236",
-year="2003",}
+year="2003",
+doi="10.1007/s00041-003-0012-2"}
 
 
 
