### 2021-06-21 #3310 @pibion
:wave: @bakerjw, @dubourg, @SebastianHoerning, @jackpiercebrown, & @song2001, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
