### 2021-06-24 #3310 @pibion
👋 @sealhuang, @Kevin-Mattheus-Moerman, @vitorsr, @stsievert, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.
