### 2021-07-01 #3310 @SebastianHoerning
Sorry, I have to decline too. I'm swamped with proposal writing at the moment.

Dr.-Ing. Sebastian Hörning

BSc, MSc, PhD

Energi Simulation Postdoctoral Research Fellow

Centre for Natural Gas

The University of Queensland

Brisbane Qld 4072 Australia

T +61 7 3346 3384  M +61 4 5114 5367

E ***@***.******@***.***>  W natural-gas.centre.uq.edu.au<https://natural-gas.centre.uq.edu.au/>

CRICOS 00025B

 [cid:e52bee4d-ff0b-4e68-9889-5b3ad47a2411]

The University of Queensland is embracing the Green Office philosophy. Please consider the environment before printing this email.

This email (including any attached files) is intended solely for the addressee and may contain confidential information of The University of Queensland. If you are not the addressee, you are notified that any transmission, distribution, printing or photocopying of this email is prohibited. If you have received this email in error, please delete and notify me. Unless explicitly stated, the opinions expressed in this email do not represent the official position of The University of Queensland.
