### 2021-08-08 #3310 @arfon
@fjprichard – looks like it's proving a little challenging to find reviewers here. Could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission?
