### 2021-10-01 #3310 @danielskatz
Sorry, my mistake.

@pibion - how do we move this forward?

@fjprichard - did you have any suggestions, as requested a couple of months ago in https://github.com/openjournals/joss-reviews/issues/3310#issuecomment-894784670?
