### 2021-10-12 #3821 @whedon
[REVIEW]: PyAFBF: a Python library to sample image textures from the anisotropic fractional Brownian field.
**Submitting author:** @fjprichard (<a href="http://orcid.org/0000-0001-5146-9894">Frédéric Richard</a>)
**Repository:** https://github.com/fjprichard/PyAFBF/
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.5
