### 2021-10-12 #3821 @pibion
@fjprichard, @arfon do you have any thoughts about why the pdf compilation is failing?  Is the pdf in a non-standard location?
