### 2021-11-04 76e124c
Update README.rst
"However, the repository documentation itself (mainly referring to README.rst) is not prominent enough to provide a good overview of the package."

I added the content of the documentation in the README file to improve the overview.
