### 2021-11-04 #3821 @fjprichard
Dear visitorsr,

Thank you for your review. Please find below some answers to your comments.

Major comments:

"I am unable to replicate the auto_examples (Example gallery) as I have not found enough information to replicate the development environment and locally generate the aforementioned documentation.

The same holds for the nose tests."

Information about the environment are in the setup file (packages required and versions). Would you suggest to have it somewhere else?

-----------

Minor comments:

_"The API is incredibly well documented. Nice-to-haves for researchers and alike such as the inclusion of a glossary are of immense help. However, the repository documentation itself (mainly referring to README.rst) is not prominent enough to provide a good overview of the package."_

I added the content of the documentation in the README file.

"Everything is otherwise stellar."
Thank you.
