### 2021-11-05 #3821 @vitorsr
Hi,

Thanks for addressing my comments.

> Information about the environment are in the setup file (packages required and versions). Would you suggest to have it somewhere else?

Yes. I am not sure if the following is clear, however please do note that the development requirements are a superset of the package requirements. The former includes packages for tests, coverage, documentation etc., to name a few development requirements. Moreover please note that I can only infer that [`nose`](https://github.com/fjprichard/PyAFBF/search?q=nose) is a development requirement because `setup.cfg` has an entry called [`nosetests`](https://github.com/fjprichard/PyAFBF/search?q=nosetests).

A common heuristic to determine precisely which are the development requirements is to match your environment from scratch, noting which packages you require to run tests, coverage, documentation etc. Commonly this information is included either in a `pip`-compatible `requirements-dev.txt` file, in a `Pipfile`, in a documentation entry addressing contributors, or with the `setuptools` `extras_requires` optional dependencies keyword.

Properly defining development requirements is useful in the open source community to open doors for contributors. As I mentioned, I cannot reproduce your environment as there are no steps on how to do so. Critically, this means I am not able to run your tests. I can verify that your code is sound, but I cannot run your test suite.
