### 2021-12-08 f6f79e2
add PyAFBF.egg-info
