+++ requirements/defaults.txt
@@ -1,0 +1,3 @@
+numpy>=1.19.2
+matplotlib>=3.3.2
+scipy>=1.5.2
