+++ requirements/defaults.txt
@@ -1,3 +1,5 @@
+--index-url https://pypi.python.org/simple/
+
 numpy>=1.19.2
 matplotlib>=3.3.2
 scipy>=1.5.2
