+++ examples/README.rst
@@ -1,5 +1,9 @@
 Example gallery
 ===============
 
+This directory contains the examples of the gallery of the [documentation](https://fjprichard.github.io/PyAFBF/auto_examples/index.html).
+These examples can be easily reproduced by downloading the python source code or jupyter notebook from the documentation.
+
+
 Basic examples
 --------------
