+++ examples/README.rst
@@ -1,7 +1,7 @@
 Example gallery
 ===============
 
-This directory contains the examples of the gallery of the (documentation)[https://fjprichard.github.io/PyAFBF/auto_examples/index.html].
+This directory contains the examples of the gallery of the   `documentation <https://fjprichard.github.io/PyAFBF/auto_examples/index.html>`_.
 These examples can be easily reproduced by downloading the python source code or jupyter notebook from the documentation.
 
 
