+++ tests/README.rst
@@ -1,3 +1,7 @@
-To test the package installation, it suffices to run the python program TestAll.py. The tests are done using doctest.
+To test the package installation, first install required packages:
 
-if the program ends up without any comment, then the package is correctly installed. Otherwise the errors are given in the delivered comments.
+pip install -r requirements/dev.txt
+
+and then execute TestAll.py.
+
+The tests are done using doctest. if the program ends up without any comment, then the package is correctly installed. Otherwise the errors are given in the delivered comments.
