### 2021-12-08 ea8de50
Update setup.cfg
