+++ setup.cfg
@@ -1,4 +1,3 @@
-[nosetests]
 verbosity=2
 where= .
 with-coverage=1
