### 2021-12-08 42463ac
Update setup.cfg
