+++ setup.cfg
@@ -1,3 +1,4 @@
+[nosetests]
 verbosity=2
 where= .
 with-coverage=1
