### 2021-12-08 #3821 @fjprichard
Dear Vitor,

Thank you for your explanation. I did some corrections on the package to improve these points:

- I added a requirement directory where I included pip compatible requirement files (defaults and dev).
- I added a line in the setup for extras_require
- I added a README file in the tests directory to explain how to run the tests with doctest.
- I added a PyAFBF.egg-info directory where such information can also be found.

I hope this will help contributors to reproduce the environment.

Best regards,
Frederic
