### 2021-12-15 #3821 @janmandel
> 👋 @janmandel, please update us on how your review is going (this is an automated reminder).

I am at the stage of checking and unchecking boxes. I did run the software back in October but I missed this issue then, any relevant messages got buried in an avalanche of notifications from this repository. Obvously I found it now and I am back at it! Sorry about that.
