### 2022-02-05 #3821 @vitorsr
Dear professor Richard,

Thanks for notifying me for I had not received your answers. I apologize if
this caused any delay. I have now updated the review checklist.

This concludes my review.

My kind regards,

Vítor

On Fri, Feb 4, 2022 at 8:59 AM Frederic Richard ***@***.***>
wrote:

> Dear @vitorsr <https://github.com/vitorsr> ,
> I wonder if you received my answers to your comments sent on december, 8th.
> Best regards,
> Frederic
>
