### 2022-02-19 #3821 @janmandel
!pip install PyAFBF
errors out on Google Colab,. To make it work,

!pip install imgaug==0.2.6
!pip install PyAFBF
