### 2022-02-21 #3821 @janmandel
I copied the examples from the documentation to a google colab notebook and they work fine.
