+++ README.rst
@@ -20,6 +20,11 @@ The package source can be downloaded from the `repository <https://github.com/fj
 The package can be installed through PYPI with
  
  pip install PyAFBF
+ 
+If the package is to be installed in a googlecollab environment, please type
+
+ !pip install imgaug==0.2.6
+ !pip install PyAFBF
 
 Communication to the author
 ===========================
