### 2022-06-07 #3821 @pibion
@arfon this paper is in a non-standard location , do you have a suggestion for how to proceed?

@fjprichard, @vitorsr pointed out:

> Yes, paper.md seems to be on a separate branch called joss, in a folder of the same name:
> https://github.com/fjprichard/PyAFBF/blob/joss/joss/paper.md

We need the paper to build, so you may need to move `paper.md` so it conforms to the expected location.
