### 2022-06-07 #3821 @pibion
@arfon on the form (https://joss.theoj.org/papers/new) it does look like you can set the branch that contains `paper.md` - can this be done after the paper is submitted?
