+++ joss/refs.bib
@@ -227,7 +227,8 @@ journal="Revista {M}athem\'atica {I}beroamericana",
 volume="13",
 number="1",
 pages="19-89",
-year="1997"}
+year="1997",
+doi="10.4171/rmi/217"}
 
 @techreport{Peltier96,
 author="Peltier, R.F. and  Levy Vehel, J.",
