+++ joss/refs.bib
@@ -30,6 +30,7 @@ edition = {2nd}
   publisher={John Wiley \& Sons}
 }
 
+
 @article{liu2019advances,
   title={Advances in Gaussian random field generation: a review},
   author={Liu, Y. and Li, J. and Sun, S. and Yu, B.},
