### 2022-06-13 4fcea50
Update paper.md
    line 9-10, the sentence that begins, "It enables to set ..." -> "Users can generate a wide variety of textures by setting the simulated model manually or randomly."
    line 10, "It also include tools" -> "It also includes tools"
    line 65, "A documentation" -> "Documentation"
