### 2022-06-13 #3821 @fjprichard
> point_rightpage_facing_up [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03821/joss.03821/10.21105.joss.03821.pdf) page_facing_up [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.03821/joss.03821/10.21105.joss.03821.pdf) page_facing_up point_left

I read the proof. It's fine.
