### 2022-06-22 #3821 @pibion
@fjprichard I have a few more notes on minor corrections to the paper - I'm working on those now and will have them up tonight.  Then I'll ask you to do a code release and generate a DOI (zenodo is a popular choice for that) and we'll recommend acceptance.

Apologies for not getting you the complete set of edits earlier, I'm moving this week and things have been hectic!
