### 2022-06-22 #3821 @pibion
- line 16, reasearchs -> reasearchs
- line 19, "Simulations it generates" -> "Its simulations"?
- line 21, "The estimation issues tackled in these works concerned the functional parameters and features of the AFBF and related models. There are still largely open.", it sounds like you're saying there are questions that are still open.  But it's not clear what those questions are.
- line 28, "among which" -> "among which are"
- lines 30, 32, no comma at the end of the list item
- line 33, "None of these package deals directly with" -> "None of these packages deal directly with"
- line 33, maybe omit the "the" in "deals directly with the anisotropic Brownian field"
- line 37, "which both differ from the one of an AFBF", I'm not sure what "the one" refers to.  Perhaps "which both differ from the anisotropy of an AFBF"?
- line 40, "The implementation of random field simulation methods in Python are less manifest" -> "There are fewer implementations of random field simulation methods in Python"?  I'm not exactly sure what you mean by "less manifest" - do you mean they're harder to find, or less developed?
- line 43, "the package AFBF" -> "the package PyAFBF", "PyAFBF" should be bold and not italicized to match the rest of the paper
- lines 44-48 is in the paragraph about Python options, but it feels like a general statement about the usefulness of AFBF simulators.  Consider moving these sentences to line 25?
- equation between lines 51, 52, is a symbol missing in |h, u>|?  Should it be |<h, u>|?
- line 53, "These functions determined" -> "These functions determine"
- line 54, "the aspect of the textures that are sampled from it" -> "the textures that are sampled from it"?
- line 56, "that enable to easily set an AFBF" -> "that enable users to easily set an AFBF"
- line 56, "in a proper way" is intriguing to me but unclear.  Are there typical mistakes or pitfalls people encounter when setting up an AFBF?  Are there a lot of parameters and it's just a pain?  I would recommend either dropping "in a proper way" or making this more specific
- line 59, "This are simulated" -> "AFBF are simulated"?
