+++ joss/paper.md
@@ -28,32 +28,30 @@ The Python library **PyAFBF** is devoted to the simulation of anisotropic textur
 
 # Statement of need
 
-For the simulation of random fields, the library **PyAFBF** relies upon the turning-band method developed in [@Bierme-2015-TBM]. This method was historically designed to facilitate researchs on the anisotropic fractional Brownian field [@Bonami2003] and related models [@Benassi97; @Guyon2000; @polisano2014texture; @Peltier96; @Vu2020]. Simulations it generates have been particularly useful to set experimental protocols aiming at assessing performances of estimation methods [@Bierme08-ESAIM; @Richard-2010; @Richard-2015; @Richard-2015b; @Richard-2016; @Richard-2017]. The estimation issues tackled in these works concerned the functional parameters and features of the AFBF and related models. There are still largely open. The library **PyAFBF** is intended to support future experiments and help mathematicians to validate and compare their methods. In particular, it could be the basis for collaborative works aiming at the development of benchmarks and data challenges.
+For the simulation of random fields, the library **PyAFBF** relies upon the turning-band method developed in [@Bierme-2015-TBM]. This method was historically designed to facilitate reasearchs on the anisotropic fractional Brownian field [@Bonami2003] and related models [@Benassi97; @Guyon2000; @polisano2014texture; @Peltier96; @Vu2020]. Its simulations have been particularly useful to set experimental protocols aiming at assessing performances of estimation methods [@Bierme08-ESAIM; @Richard-2010; @Richard-2015; @Richard-2015b; @Richard-2016; @Richard-2017]. The issues tackled in these works concerned the statistical estimation of these functional parameters and features of the AFBF and related models. There are still largely open. The library **PyAFBF** is intended to support future experiments and help mathematicians to validate and compare their methods. In particular, it could be the basis for collaborative works aiming at the development of benchmarks and data challenges. The library is also of interest for researchers in image processing where random fields can serve as texture or noise models for medical images [@Bierme09-ESAIM; @Bierme10-Springer; @Richard-2010; @Richard-2015b; @Richard-2016] or photographic films [@Richard-2017]. This is also of interest for researchers in machine learning where the random field simulation could be included in image generative models such as GAN.
 
-The simulation of random fields is a classical topic of spatial statistics [@lantuejoul2013geostatistical; @cressie2015statistics; @Chiles-2012]. As reviewed in [@liu2019advances], there are some R packages devoted to this topic, among which:
+The simulation of random fields is a classical topic of spatial statistics [@lantuejoul2013geostatistical; @cressie2015statistics; @Chiles-2012]. As reviewed in [@liu2019advances], there are some R packages devoted to this topic, among which are:
 
-- [RandomFields](https://cran.r-project.org/web/packages/RandomFields/) [@Schlater-2015] that enables the simulation of stationary fields and also some non-stationary Gaussian random fields such max-stable fields,
+- [RandomFields](https://cran.r-project.org/web/packages/RandomFields/) [@Schlater-2015] that enables the simulation of stationary fields and also some non-stationary Gaussian random fields such max-stable fields
     
-- [FieldSim](https://cran.r-project.org/web/packages/FieldSim/FieldSim.pdf) [@brouste2007fractional] that allows the simulation of manifold indexed Gaussian field,
+- [FieldSim](https://cran.r-project.org/web/packages/FieldSim/FieldSim.pdf) [@brouste2007fractional] that allows the simulation of manifold indexed Gaussian field
     
 
[…]
