### 2022-06-23 #3821 @fjprichard
@pibion I created a new release (v0.1.0) for the package including the joss paper. I have a question about the DOI, just to be sure. The DOI is for the new release ?  and not for the paper.
