+++ afbf/__init__.py
@@ -5,7 +5,7 @@ __author__ = "Frédéric Richard."
 __copyright__ = "2021."
 __credits__ = ["Frédéric Richard"]
 __license__ = "GNU GENERAL PUBLIC LICENSE, Version 3."
-__version__ = "0.1.1"
+__version__ = "0.1.3"
 __maintainer__ = "Frédéric Richard"
 __email__ = "frederic.richard_at_univ-amu.fr"
 __status__ = "Release"
