### 2022-06-23 c1b4038
Update README.rst
add a badge for DOI
