### 2022-06-23 #3821 @fjprichard
@pibion I created the doi. The badge of the DOI is included in the README file of the new release (v0.1.3). Thank you very much.
