### 2022-06-29 #3821 @pibion
Hi @fjprichard , thank you for your edits!  I have one outstanding minor comment and two more involved comments.  I've tried to suggest text for my involved comments.

Apologies for all the nitpickiness!  I promise this is at least somewhat useful, as it will make the process faster once I recommend your paper for publication.  The JOSS release managers will ask for edits if anything in the paper is unclear.

### A minor comment
line 16, reasearchs -> reasearchs

### PyAFBF supports investigation of open questions about anisotropic fields
The sentences that start on line 19 are still unclear to me.  Currently your paper says:

> Its simulations have been particularly useful to set experimental protocols aiming at assessing performances of estimation methods (citations). The issues tackled in these works concerned the statistical estimation of these functional parameters and features of the AFBF and related models. There are still largely open.

I'm not sure what "functional parameters" refers to.  I also get the sense that you're saying that there are still open questions, but I don't know what those questions are based on your text.

After reading through some of your cited papers, here's my suggestion.  Is this what you're saying, or am I missing important points?

> Its simulations have been particularly useful for testing estimators designed to identify fields as isotropic or anisotropic (citations).  We still do not have estimators that reliably discriminate isotropic from anisotropic fields for all possible field parameters.

### General to specific in the statement of need

I would recommend moving the following to the very beginning of the "Statement of Need" section as it is the most general information:

> The library is also of interest for researchers in image processing where random fields can serve as texture or noise models for medical images (Biermé et al., 2009; Biermé & Richard, 2011; Richard, 2015, 2016a; Richard & Biermé, 2010) or photographic films (Richard, 2017). This is also of interest for researchers in machine learning where the random field simulation could be included in image generative models such as GAN.

I also have the following suggestions to make it read a little smoother.  Is this missing important points, or oversimplifying?  I think it is worth explicity saying that there is a need for reliable estimators that accurately identify anisotropic fields, this makes it very clear that your library is needed for testing metrics.

> Simulating anisotropic fields is useful both for developing metrics that identify anisotropic fields and for generative methods that need anisotropic fields as input.  Isotropy metrics have applications from medical diagnostics to biometric authentication (Richard, F. J. P. (2016b)).  Anisotropic field samples are needed for  texture or noise models for medical images (Biermé et al., 2009; Biermé & Richard, 2011; Richard, 2015, 2016a; Richard & Biermé, 2010), photographic films (Richard, 2017), and other image-generative applications such as Generative Adversarial Networks.
