### 2022-07-01 1e43ec1
Update paper.md
Response to the review of 06/29/2022.
