### 2022-07-01 #3821 @fjprichard
@pibion Thank you for your valuable comments; there is no problem about continuing improving the text. I rewrite the introduction to take into account your comments. I separated it into two paragraphs: the first one presents generic interests of the package. The second one is more focused on its interest for the evaluation of estimation methods. In this part, I tried to be more specific about the estimation issues. I hope it is "smoother".
