### 2022-07-07 #3821 @pibion
- line 16, reasearchs -> reasearchs
- line 31, developping -> developing

Simulations of AFBF have been particularly useful to evaluate methods devoted to the estimation of the features derived from the parameters of the AFBF, namely the topothesy and the Hurst functions.  Such features are important to characterize directional properties of the field and describe the image texture anisotropy. They can be use to classify or segment images, and detect image abnormalities. However, issues raised by the statistical estimation of the model parameters and features remain open.

> Simulations of AFBF have been particularly useful to evaluate methods devoted to the estimation of the features derived from the parameters of the AFBF, namely the topothesy and the Hurst functions.

I want to check that I'm understanding correctly: Are the topothesy and Hurst functions the "methods that estimate the features derived from the parameters of the AFBF," or are they the "features derived from the parameters of the AFBF?"

> Such features are important to characterize directional properties of the field and describe the image texture anisotropy. They can be use to classify or segment images, and detect image abnormalities.

I'm not confident I know what "features" refers to - the topothesy and Hurst functions?  Features of the AFBF like directionality?  Statistically-calculated estimators?  If "features" really means the topothesy and Hurst functions, maybe "Such functions(?) help characterize directional properties of the field and describe the image texture anisotropy."

> They can be use to classify or segment images, and detect image abnormalities.

This is wonderfully clear, no comments :)

> However, issues raised by the statistical estimation of the model parameters and features remain open.

I don't have any sense of what "issues raised" could be.  I think this would be clearer if you gave a concrete example of an open issue.  Like, "However, we do not have accurate statistical estimators for model parameters and features for many AFBFs."
