+++ joss/paper.md
@@ -30,7 +30,9 @@ The Python library **PyAFBF** is devoted to the simulation of anisotropic textur
 
 For the simulation of random fields, the library **PyAFBF** relies upon the turning-band method developed in [@Bierme-2015-TBM]. This method was historically designed to facilitate reasearchs on the anisotropic fractional Brownian field (AFBF) [@Bonami2003] and related models [@Benassi97; @Guyon2000; @polisano2014texture; @Peltier96; @Vu2020]. The library is of interest for researchers in image processing where random fields can serve as texture or noise models for medical images [@Bierme09-ESAIM; @Bierme10-Springer; @Richard-2010; @Richard-2015b; @Richard-2016] or photographic films [@Richard-2017]. It could also be interesting for machine learning researchers who could include the random field simulation in the design or the learning of image generative models such as Generative Adversarial Networks. 
 
-Besides, simulations of AFBF have been particularly useful to evaluate methods devoted to the estimation of the features derived from the parameters of the AFBF, namely the topothesy and the Hurst functions [@Bierme08-ESAIM; @Richard-2010; @Richard-2015; @Richard-2015b; @Richard-2016; @Richard-2017]. Such features are important to characterize directional properties of the field and describe the image texture anisotropy. They can be use to classify or segment images, and detect image abnormalities. However, issues raised by the statistical estimation of the model parameters and features remain open. Hence, while developing new estimation methods, mathematicians could use the library **PyAFBF** to design experiments and assess performances of their methods. In particular, the library could be the basis for collaborative works aiming at the development of benchmarks and data challenges concerning the estimation issues. 
+Besides, simulations of AFBF have been particularly useful to evaluate methods devoted to the estimation of the features derived from the parameters of the AFBF [@Bierme08-ESAIM; @Richard-2010; @Richard-2015; @Richard-2015b; @Richard-2016; @Richard-2017]; see below for the definition of these parameters called the topothesy and Hurst functions. Such features are important to characterize directional properties of the field and describe the image texture anisotropy. They can be use to classify or segment images, and detect image abnormalities. 
+
+However, the statistical estimation of the model parameters and features remains an open issue. For instance, the methods proposed for the estimation of the topothesy and Hurst functions are incomplete: in [@Richard-2016], it only estimates the topothesy function in directions where the Hurst function is minimal. In [@Bierme08-ESAIM], it can only be applied to the estimation of the Hurst function in a few directions. Hence, for the development of new estimation methods, mathematicians could use the library **PyAFBF** to design experiments and assess performances of their methods. In particular, the library could be the basis for collaborative works aiming at the development of benchmarks and data challenges concerning the estimation issues. 
 
 The simulation of random fields is a classical topic of spatial statistics [@lantuejoul2013geostatistical; @cressie2015statistics; @Chiles-2012]. As reviewed in [@liu2019advances], there are some R packages devoted to this topic, among which are:
 
