+++ joss/paper.md
@@ -32,7 +32,7 @@ For the simulation of random fields, the library **PyAFBF** relies upon the turn
 
 Besides, simulations of AFBF have been particularly useful to evaluate methods devoted to the estimation of the features derived from the parameters of the AFBF [@Bierme08-ESAIM; @Richard-2010; @Richard-2015; @Richard-2015b; @Richard-2016; @Richard-2017]; see below for the definition of these parameters called the topothesy and Hurst functions. Such features are important to characterize directional properties of the field and describe the image texture anisotropy. They can be use to classify or segment images, and detect image abnormalities. 
 
-However, the statistical estimation of the model parameters and features remains an open issue. For instance, the methods proposed for the estimation of the topothesy and Hurst functions are incomplete: in [@Richard-2016], it only estimates the topothesy function in directions where the Hurst function is minimal. In [@Bierme08-ESAIM], it can only be applied to the estimation of the Hurst function in a few directions. Hence, for the development of new estimation methods, mathematicians could use the library **PyAFBF** to design experiments and assess performances of their methods. In particular, the library could be the basis for collaborative works aiming at the development of benchmarks and data challenges concerning the estimation issues. 
+However, the statistical estimation of the model parameters and features still raises open issues. For instance, the actual methods for the estimation of the topothesy and Hurst functions are incomplete: in [@Richard-2017], the method only estimates the topothesy function in directions where the Hurst function is minimal. In [@Bierme08-ESAIM], it can only be used for the estimation of the Hurst function in the horizontal and vertical directions. Hence, for the development of new estimation methods, mathematicians could use the library **PyAFBF** to design experiments and assess performances of their methods. In particular, the library could be the basis for collaborative works aiming at the development of benchmarks and data challenges concerning the estimation issues. 
 
 The simulation of random fields is a classical topic of spatial statistics [@lantuejoul2013geostatistical; @cressie2015statistics; @Chiles-2012]. As reviewed in [@liu2019advances], there are some R packages devoted to this topic, among which are:
 
