### 2022-07-18 59514f0
Update paper.md
Answers to the review (07/07/2022).
