+++ joss/paper.md
@@ -28,11 +28,11 @@ The Python library **PyAFBF** is devoted to the simulation of anisotropic textur
 
 # Statement of need
 
-For the simulation of random fields, the library **PyAFBF** relies upon the turning-band method developed in [@Bierme-2015-TBM]. This method was historically designed to facilitate reasearchs on the anisotropic fractional Brownian field (AFBF) [@Bonami2003] and related models [@Benassi97; @Guyon2000; @polisano2014texture; @Peltier96; @Vu2020]. The library is of interest for researchers in image processing where random fields can serve as texture or noise models for medical images [@Bierme09-ESAIM; @Bierme10-Springer; @Richard-2010; @Richard-2015b; @Richard-2016] or photographic films [@Richard-2017]. It could also be interesting for machine learning researchers who could include the random field simulation in the design or the learning of image generative models such as Generative Adversarial Networks. 
+For the simulation of random fields, the library **PyAFBF** relies upon the turning-band method developed in [@Bierme-2015-TBM]. This method was historically designed to facilitate researchs on the anisotropic fractional Brownian field (AFBF) [@Bonami2003] and related models [@Benassi97; @Guyon2000; @polisano2014texture; @Peltier96; @Vu2020]. The library is of interest for researchers in image processing where random fields can serve as texture or noise models for medical images [@Bierme09-ESAIM; @Bierme10-Springer; @Richard-2010; @Richard-2015b; @Richard-2016] or photographic films [@Richard-2017]. It could also be interesting for machine learning researchers who could include the random field simulation in the design or the learning of image generative models such as Generative Adversarial Networks. 
 
-Besides, simulations of AFBF have been particularly useful to evaluate methods devoted to the estimation of the features derived from the parameters of the AFBF [@Bierme08-ESAIM; @Richard-2010; @Richard-2015; @Richard-2015b; @Richard-2016; @Richard-2017]; see below for the definition of these parameters called the topothesy and Hurst functions. Such features are important to characterize directional properties of the field and describe the image texture anisotropy. They can be used to classify or segment images, and detect image abnormalities. 
+Besides, AFBF are characterized by two parameters called the topothesy function and the Hurst functions; see below for a definition. These parameters, as well as derived features, are important to characterize directional properties of the field and describe the image texture anisotropy. They can be used to classify or segment images, and detect image abnormalities. Achieving such tasks requires estimating parameters and features. However the statistical estimation of AFBF features is still an open issue. In particular, actual methods for the estimation of the topothesy and Hurst functions are incomplete: the method proposed in [@Richard-2017] only estimates the topothesy function in directions where the Hurst function is minimal. The method in [@Bierme08-ESAIM] can only be used for the estimation of the Hurst function in the horizontal and vertical directions.
 
-However, the statistical estimation of the model parameters and features still raises open issues. For instance, the actual methods for the estimation of the topothesy and Hurst functions are incomplete: in [@Richard-2017], the method only estimates the topothesy function in directions where the Hurst function is minimal. In [@Bierme08-ESAIM], it can only be used for the estimation of the Hurst function in the horizontal and vertical directions. Hence, for the development of new estimation methods, mathematicians could use the library **PyAFBF** to design experiments and assess performances of their methods. In particular, the library could be the basis for collaborative works aiming at the development of benchmarks and data challenges concerning the estimation issues. 
[…]
