+++ joss/paper.md
@@ -30,9 +30,9 @@ The Python library **PyAFBF** is devoted to the simulation of anisotropic textur
 
 For the simulation of random fields, the library **PyAFBF** relies upon the turning-band method developed in [@Bierme-2015-TBM]. This method was historically designed to facilitate researchs on the anisotropic fractional Brownian field (AFBF) [@Bonami2003] and related models [@Benassi97; @Guyon2000; @polisano2014texture; @Peltier96; @Vu2020]. The library is of interest for researchers in image processing where random fields can serve as texture or noise models for medical images [@Bierme09-ESAIM; @Bierme10-Springer; @Richard-2010; @Richard-2015b; @Richard-2016] or photographic films [@Richard-2017]. It could also be interesting for machine learning researchers who could include the random field simulation in the design or the learning of image generative models such as Generative Adversarial Networks. 
 
-Besides, AFBF are characterized by two parameters called the topothesy function and the Hurst functions; see below for a definition. These parameters, as well as derived features, are important to characterize directional properties of the field and describe the image texture anisotropy. They can be used to classify or segment images, and detect image abnormalities. Achieving such tasks requires estimating parameters and features. However the statistical estimation of AFBF features is still an open issue. In particular, actual methods for the estimation of the topothesy and Hurst functions are incomplete: the method proposed in [@Richard-2017] only estimates the topothesy function in directions where the Hurst function is minimal. The method in [@Bierme08-ESAIM] can only be used for the estimation of the Hurst function in the horizontal and vertical directions.
+Besides, AFBF are characterized by two parameters called the topothesy function and the Hurst functions; see below for a definition. These parameters, as well as derived features, are important to characterize directional properties of the field and describe the image texture anisotropy. They can be used to classify or segment images, and detect image abnormalities. Achieving such tasks requires estimating parameters and features. However the statistical estimation of AFBF features is still a topic of research. In particular, actual methods for the estimation of the topothesy and Hurst functions are incomplete: the method proposed in [@Richard-2017] only estimates the topothesy function in directions where the Hurst function is minimal. The method in [@Bierme08-ESAIM] can only be used for the estimation of the Hurst function in the horizontal and vertical directions. 
 
-In a research context, simulations of AFBF have been particularly useful for the evaluation of estimation methods [@Bierme08-ESAIM; @Richard-2010; @Richard-2015; @Richard-2015b; @Richard-2016; @Richard-2017]. Hence, in the future, the library **PyAFBF** could be used by mathematicians to design experiments and assess method performances. In particular, the library could be the basis for collaborative works aiming at the development of benchmarks and data challenges concerning the estimation issue. 
+In this research context, simulations of AFBF have been particularly useful for the evaluation of estimation methods [@Bierme08-ESAIM; @Richard-2010; @Richard-2015; @Richard-2015b; @Richard-2016; @Richard-2017]. Hence, in the future, the library **PyAFBF** could be used by mathematicians to design experiments and assess method performances. In particular, the library could be the basis for collaborative works aiming at the development of benchmarks and data challenges concerning the estimation issue. 
 
 The simulation of random fields is a classical topic of spatial statistics [@lantuejoul2013geostatistical; @cressie2015statistics; @Chiles-2012]. As reviewed in [@liu2019advances], there are some R packages devoted to this topic, among which are:
 
