### 2022-07-18 #3821 @fjprichard
**line 16, reasearchs -> researchs
    line 31, developping -> developing**

Done.

**I want to check that I'm understanding correctly: Are the topothesy and Hurst functions the "methods that estimate the features derived from the parameters of the AFBF," or are they the "features derived from the parameters of the AFBF?"
I'm not confident I know what "features" refers to - the topothesy and Hurst functions? Features of the AFBF like directionality? Statistically-calculated estimators? If "features" really means the topothesy and Hurst functions, maybe "Such functions(?) help characterize directional properties of the field and describe the image texture anisotropy."**

The topothesy function and Hurst functions are the two parameters of the field (they are defined after the first equation). We can define features

**However, issues raised by the statistical estimation of the model parameters and features remain open.
I don't have any sense of what "issues raised" could be. I think this would be clearer if you gave a concrete example of an issue. Like, "However, we do not have accurate statistical estimators for model parameters and features for many AFBFs."**

I modified the text to be more specific.
