### 2022-07-18 #3821 @pibion
@fjprichard this version is very clear, thank you!  My only comment is that on line 16 I think that "researchs" should be "research".

Once you've made that change, make a tagged release and archive, and report the version number and archive DOI in this thread.  Then I'll follow the steps at https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance, which end in recommending acceptance!

We're almost there!
